Organize hardware modules into dedicated package
