+++ OcchioOnniveggente/src/hardware/audio.py
@@ -448,7 +448,7 @@ def create_local_tts(container: "ServiceContainer") -> TextToSpeech:
 
 
 try:  # pragma: no cover - registry may not be available during docs build
-    from .plugins import register_stt, register_tts
+    from ..plugins import register_stt, register_tts
 
     register_stt("local", create_local_stt)
     register_tts("local", create_local_tts)
