+++ OcchioOnniveggente/src/main.py
@@ -23,8 +23,8 @@ import logging
 
 from src.config import Settings, get_openai_api_key
 from src.filters import ProfanityFilter
-from src.audio import AudioPreprocessor, record_until_silence, play_and_pulse
-from src.lights import SacnLight, WledLight, color_from_text
+from src.hardware.audio import AudioPreprocessor, record_until_silence, play_and_pulse
+from src.hardware.lights import SacnLight, WledLight, color_from_text
 from src.oracle import (
     transcribe,
     fast_transcribe,
@@ -42,7 +42,7 @@ from src.conversation import ConversationManager, DialogState
 from src.logging_utils import setup_logging
 from src.language_session import update_language
 from src.cli import _ensure_utf8_stdout, say, oracle_greeting, default_response
-from src.audio_device import pick_device, debug_print_devices
+from src.hardware.audio_device import pick_device, debug_print_devices
 from src.profile_utils import get_active_profile, make_domain_settings
 
 logger = logging.getLogger(__name__)
