+++ OcchioOnniveggente/src/realtime_oracolo.py
@@ -28,7 +28,7 @@ import sounddevice as sd
 import websockets
 from src.conversation import ConversationManager, DialogState
 from src.retrieval import load_questions
-from src import local_audio
+from src.hardware import local_audio
 from src.oracle import (
     enqueue_generate_reply,
     enqueue_synthesize_voice,
