+++ docs/tutorials/stt_tts.md
@@ -3,7 +3,7 @@
 Esempio minimale per usare i modelli locali:
 
 ```python
-from OcchioOnniveggente.src.local_audio import transcribe, synthesize
+from OcchioOnniveggente.src.hardware.local_audio import transcribe, synthesize
 
 text = transcribe("audio.wav")
 print(text)
